requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
sentence-transformers
faiss-cpu
numpy
python-dotenv
requests
